Fix prober segfault (#11)
* Fix prober segfault

When a connection error occurred, the prober printed the error, but
continued to try to get the status code of a nil object. This change
logs an error (now to stderr) but does not continue. We now count the
number of failures (which should be zero, but we are seeing Fulcio
returning failures frequently).

Also removed some of the other logging unless there was an unexpected
value.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Add worker pool

This limits the number of concurrent requests to prevent connection
errors.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>