fix: add a timeout of 10 minutes to the probers
There is not timeout set for those jobs, and by default
github sets it to 360 minutes. It means that any step
that is stuck will cause this to run for 6 hours.

Adding a more sane timeout of 10 minutes to ensure that
the job is completed on time, or abort if not.

Closes: #1398
Signed-off-by: Yolanda Robla <info@ysoft.biz>