Create github issues for alerts
Add a job to the reusable prober workflow to create an issue in
sigstore/public-good-instance when one of the probes fails.

The new job needs a new input, SIGSTORE_BOT_PAT, configured as a secret
for the sigstore/sigstore-probers repo. The PAT needs to belong to a
user or bot that has permission to create issues with labels for
sigstore/public-good-instance.

The new job creates the issue with a new `needs-assignee` label. A
sibling job in sigstore/public-good-instance will react to new issues
that have this label and add the active oncall engineer as the assignee.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>